+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Default zoom on GPS location changed from level 19 to 18
+- Search for the route target will now use the origin as reference point
 
 ### Fixed